+++ payments/src/sdk/models/shared/memberclient.ts
@@ -16,6 +16,11 @@ import { Threed33fba3f009de957b3be92fba006d6383af7e39f823cc1fd213506f6205100f } 
 import { UserRoles } from "./userroles";
 import { Expose, Type } from "class-transformer";
 
+export enum ClientUserAccountType {
+    Member = "member",
+    Enterprise = "enterprise",
+}
+
 export enum MemberClientStatus {
     Active = "Active",
     Inactive = "Inactive",
@@ -48,6 +53,10 @@ export class MemberClient extends SpeakeasyBase {
     @SpeakeasyMetadata()
     @Expose({ name: "clientId" })
     clientId: string;
+
+    @SpeakeasyMetadata()
+    @Expose({ name: "clientUserAccountType" })
+    clientUserAccountType: ClientUserAccountType;
 
     @SpeakeasyMetadata()
     @Expose({ name: "collaboratorGroupId" })